requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
flask==3.0.0
